feat: merge analyze to one page
