+++ src/routes/resources/analyze/[id]/index.tsx
@@ -1,0 +1,69 @@
+import { component$ } from '@builder.io/qwik';
+import { routeLoader$ } from '@builder.io/qwik-city';
+
+import Analyze from '~/components/analyze/Analyze';
+import analyzeProfile from '~/util/analyze/functions/analyzeProfile';
+import analyzeTimings from '~/util/analyze/functions/analyzeTimings';
+import { collector } from '~/util/analyze/functions/collector';
+import { defaultDescription, generateHead } from '~/root';
+
+export const useResults = routeLoader$(async ({ params }) => {
+
+  // paper timings id is 32 characters and spark profile id is 10 characters
+  // just in case spark decides to use more than 10, just check if it's less than 30 characters idk
+  if (params.id.length < 30) {
+    const results = await analyzeProfile(params.id);
+    try {
+      await collector(params.id, 'https://api.profiler.birdflop.com', 'spark');
+    } catch (error) {
+      console.error('Collector error:', error);
+    }
+    return results;
+  } else {
+    const results = await analyzeTimings(params.id);
+    try {
+      await collector(params.id, 'https://api.profiler.birdflop.com', 'timings');
+    } catch (error) {
+      console.error('Collector error:', error);
+    }
+    return results;
+  }
+});
+
+export default component$(() => {
+  const results = useResults();
+
+  return (
+    <Analyze>
+      <div class="w-full my-12 grid grid-cols-3 gap-4">
+        {results.value.map((field: Field, i: number) => {
+          return (
+            <div class="lum-card" key={`field${i}`}>
+              <p class="font-bold text-xl wrap-break-word">
+                {field.name.replace(/\./g, '\n> ')}
+              </p>
+              <p class="lum-text-secondary">
+                {field.value}
+              </p>
+              {field.buttons?.map((button: any, i2: number) => {
+                return (
+                  <a class="lum-btn" key={`button${i2}-${i}`} href={button.url}>
+                    {button.text}
+                  </a>
+                );
+              })}
+            </div>
+          );
+        })}
+      </div>
+      <p class="text-lum-text-secondary">
+        Got another Spark Profile to analyze?
+      </p>
+    </Analyze>
+  );
+});
+
+export const head = generateHead({
+  title: 'Automatic Minecraft Spark Profile and Timings Analyzer - Birdflop',
+  description: 'Analyze your Spark Profile and Paper Timings to get optimization recommendations. Developed by Birdflop. ' + defaultDescription,
+});
